feat(SmtForest): Create the storage abstraction
This commit contains the `Storage` and `StoredTreeHandle` traits that
encapsulate the operations that must be provided by a storage backend.
They are designed so that the backend provides simple data operations,
while the actual logic resides in the forest itself. This gives possible
implementations enough leeway to be efficient while also providing the
necessary functionality.

With no concrete implementations included in this commit, this serves
primarily as documentation of the design space. The first concrete
implementation of `Storage` and `StoredTreeHandle` will come in the
commit that contains the `SmtForest` type itself.