feat(SmtForest) Introduce the new forest API
This commit is predominantly a skeleton of the new forest API, though it
does also introduce the new `ForestOp`, `TreeBatch` and `ForestBatch`
types to manage batches of operations.

The API as it exists today is purely an expansion of the existing
`SmtForest` public API, and is compatible except for the specific error
type that is being used in `Result`-returning methods. The `MerkleError`
that is used in `SmtForest` is now a single variant of the larger error
enum.