feat(SmtForest) Implement in-memory storage for the new forest
This commit implements the `InMemoryStorage` for the new SMT forest,
allowing it to achieve feature parity with the existing `SmtForest`
implementation, though not yet providing any of the more advanced and
persistent features.

The reader will note that there are many portions of the code that look
very similar to existing structures. This is known and intentional at
this stage as it provides the means to iterate and experiment without
breaking existing code. If possible, these things will be harmonized in
the future.